Rename app_mock for current branch
